+++ src/DatabaseDefinition.php
@@ -101,7 +101,7 @@ class DatabaseDefinition
     /**
      * Specify the seeders to run.
      *
-     * @param string[] $seeders The seeders to run after migrating.
+     * @param mixed[] $seeders The seeders to run after migrating.
      * @return static
      */
     public function seeders($seeders): self